Changes to Systems
My main goal for this week was to fix the Zombie getting stuck after attacking the player bug. I also moved the player moving conditions into a method in the MovingObject class. I did this to stop clutter in the Main Class.  I was able to make the characters animate while they are moving, however, I wasn’t able to change the image of the character. I revamped the death mechanics of the game and made the zombies do a lot more damage. This was done by adding an area of effect sort of thing, where, if you are too close to a zombie, they will deal damage to the player. Previously, the only way that damage was dealt was by if the zombie and main character were touching. I wasn’t able to fix the zombie getting stuck glitch.